Misc code cleanup

+++ KernelLand/Kernel/drv/proc.c
@@ -250,8 +250,9 @@ int SysFS_UpdateFile(int ID, const char *Data, int Length)
 	for( ent = gSysFS_FileList; ent; ent = ent->Next )
 	{
 		// It's a reverse sorted list
-		if(ent->Node.Inode < ID)	return 0;
-		if(ent->Node.Inode == ID)
+		if(ent->Node.Inode < (Uint64)ID)
+			return 0;
+		if(ent->Node.Inode == (Uint64)ID)
 		{
 			ent->Node.ImplPtr = (void*)Data;
 			ent->Node.Size = Length;
@@ -278,8 +279,8 @@ int SysFS_RemoveFile(int ID)
 	for( ent = gSysFS_FileList; ent; prev = ent, ent = ent->Next )
 	{
 		// It's a reverse sorted list
-		if(ent->Node.Inode < ID)	return 0;
-		if(ent->Node.Inode == ID)	break;
+		if(ent->Node.Inode < (Uint64)ID)	return 0;
+		if(ent->Node.Inode == (Uint64)ID)	break;
 	}
 	
 	if(!ent)	return 0;
@@ -326,7 +327,7 @@ int SysFS_RemoveFile(int ID)
 char *SysFS_Comm_ReadDir(tVFS_Node *Node, int Pos)
 {
 	tSysFS_Ent	*child = (tSysFS_Ent*)Node->ImplPtr;
-	if(Pos < 0 || Pos >= Node->Size)	return NULL;
+	if(Pos < 0 || (Uint64)Pos >= Node->Size)	return NULL;
 	
 	for( ; child; child = child->Next, Pos-- )
 	{
