+++ KernelLand/Kernel/vfs/dir.c
@@ -163,7 +163,7 @@ int VFS_ReadDir(int FD, char *Dest)
 		return 0;
 	}
 	
-	if(h->Node->Size != -1 && h->Position >= h->Node->Size) {
+	if(h->Node->Size != (Uint64)-1 && h->Position >= h->Node->Size) {
 		//LEAVE('i', 0);
 		return 0;
 	}
